refactor: Consolidate ReviewsScreen components
Moves all composables from `presentation/src/main/java/com/london/presentation/feature/reviews/components` into `ReviewsScreen.kt`.
